Extract `PaymentView` to own file
